chore: bump CML to 3.1.1 & setup test extension to use wasm
There are now 2 webpack configs:
- webpack.config.sw.js is used only for service worker script
- webpack.config.js is used for everything else

UI script now has to be loaded asynchronously.
As a result, web-extension tests had to be adjusted.
