+++ packages/core/package.json
@@ -63,8 +63,7 @@
   "dependencies": {
     "@cardano-ogmios/client": "5.5.7",
     "@cardano-sdk/util": "^0.6.0",
-    "@dcspark/cardano-multiplatform-lib-browser": "^3.1.0",
-    "@dcspark/cardano-multiplatform-lib-nodejs": "^3.1.0",
+    "@dcspark/cardano-multiplatform-lib-nodejs": "^3.1.1",
     "@emurgo/cip14-js": "^3.0.1",
     "bech32": "^2.0.0",
     "lodash": "^4.17.21",
