+++ packages/e2e/test/web-extension/extension/ui-entry.ts
@@ -1,0 +1,1 @@
+import('./ui');
